added mnogo

+++ Assets/Scripts/Player/Parameters.cs
@@ -6,7 +6,7 @@ public static class Parameters
 {
     public static int power = 1;
 
-    public static int powerGrowthMultiplier = 1;
+    public static float powerGrowthMultiplier = 1;
 
     public static int money = 0;
 }
